requirements.txt
flask
flask-sqlalchemy
flask-jwt-extended
flask-migrate
python-dotenv
flask-cors
razorpay
werkzeug